change uint to size_t and some fixes in frame_aio

+++ branches/work/frame/aio/aioselector.hpp
@@ -64,7 +64,7 @@ private:
 	void doUnregisterObject(Object &_robj, int _lastfailpos = -1);
 	ulong doIo(Socket &_rsock, ulong _evs, ulong _filter = 0);
 	ulong doExecute(const ulong _pos);
-	void doPrepareObjectWait(const ulong _pos, const TimeSpec &_timepos);
+	void doPrepareObjectWait(const size_t _pos, const TimeSpec &_timepos);
 private://data
 	struct Data;
 	Data	&d;
