+++ branches/work/frame/aio/src/aioselector_epoll.cpp
@@ -157,21 +157,11 @@ int Selector::Data::computeWaitTimeout()const{
 }
 void Selector::Data::addNewSocket(){
 	++socksz;
-// 	if(socksz > sockcp){
-// 		uint oldcp = sockcp;
-// 		sockcp += 64;//TODO: improve!!
-// 		epoll_event *pevs = new epoll_event[sockcp];
-// 		memcpy(pevs, events, oldcp * sizeof(epoll_event));
-// 		delete []events;
-// 		for(uint i = 0; i < sockcp; ++i){
-// 			pevs[i].events = 0;
-// 			pevs[i].data.u64 = 0;
-// 		}
-//	}
 }
 inline epoll_event* Selector::Data::eventPrepare(
 	epoll_event &_ev, const uint32 _objpos, const uint32 _sockpos
 ){
+	//TODO:we are now limited to uint32 objects and uint32 sockets per object
 	_ev.data.u64 = _objpos;
 	_ev.data.u64 <<= 32;
 	_ev.data.u64 |= _sockpos;
@@ -705,12 +695,12 @@ ulong Selector::doExecute(const ulong _pos){
 	}
 	return rv;
 }
-void Selector::doPrepareObjectWait(const ulong _pos, const TimeSpec &_timepos){
-	Stub &stub(d.stubs[_pos]);
-	const int32 * const pend(stub.objptr->reqpos);
-	bool mustwait = true;
+void Selector::doPrepareObjectWait(const size_t _pos, const TimeSpec &_timepos){
+	Stub 					&stub(d.stubs[_pos]);
+	const size_t * const	pend(stub.objptr->reqpos);
+	bool 					mustwait = true;
 	vdbgx(Debug::aio, "stub "<<_pos);
-	for(const int32 *pit(stub.objptr->reqbeg); pit != pend; ++pit){
+	for(const size_t *pit(stub.objptr->reqbeg); pit != pend; ++pit){
 		Object::SocketStub &sockstub(stub.objptr->pstubs[*pit]);
 		//sockstub.chnevents = 0;
 		const uint8 reqtp = sockstub.requesttype;
